firefighter-server/requirements.txt
# Flask + Socket.IO
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Vector Database
qdrant-client>=1.7.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# Production server
gunicorn>=21.0.0
eventlet>=0.33.0
